Require userVerification on sign-up and login
We check that user verification flag is set for the returned credential but didn't require it on credential creation. Because the default for user verification is preferred, it wouldn't be used when for example the lid of the laptop is closed, because the finger print is not available. Now that we set user verification to required, it will prompt for the system password instead when the lid is closed.